Remove redundant awaits and cast stream data
Remove unnecessary await when delegating to a synchronous onBlockProcessed handler in PoC.ts and delete related explanatory comments about background chaining. Also cast StreamMessageEvent.data to Uint8Array in ReusableStream.ts to satisfy the handler's expected type. These are primarily cleanup/typing fixes with no behavioral change.

+++ src/src/poc/networking/stream/ReusableStream.ts
@@ -81,7 +81,7 @@ export class ReusableStream extends Logger {
 
         // Set up event listeners for the MessageStream interface
         this.messageHandler = (event: StreamMessageEvent) => {
-            void this.handleMessage(event.data);
+            void this.handleMessage(event.data as Uint8Array);
         };
 
         this.closeHandler = (_event: StreamCloseEvent) => {
